backend/pyproject.toml
[project]
name = "trilingual-customer-service"
version = "0.1.0"
description = "Trilingual customer service multi-agent system using Vertex AI ADK"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
